nodejs: fix cross compilation for armv7l
The v8 library supports only limited number of build platforms based on
the host architecture. The rule there is the bitness (the mixing of
32bit and 64bit architectures seems to be in most cases disallowed).
Thus this adds usage of the emulator of the host platform and builds
tools for that.

This also allows us to use ninja instead of make when build ≠ host while
waiting for https://github.com/nodejs/gyp-next/pull/185 to be merged.

Co-authored-by: Ivan Trubach <mr.trubach@icloud.com>